VIX-2120
Added logic to support skinny display elements like an arch.  This required adding logic to avoid a division by zero.

Fixed logic to initialize the parent property of line points so that everything refreshes after an edit in the point grid.

Added logic to prevent lasso from leaving the canvas.

Added logic so that the PolygonEditorViewModel is given the width and height of the canvas instead of using the MouseEnter event.

Updated lines so that the green point will turn pink when selected.

Added logic so that if the display element is really skinny like an arch the default shape is a line instead of a polygon.