VIX-2120
Added logic to prevent deleting polygon points when the morph polygon has a wipe fill mode.

Added logic to hide the green line segment once a polygon point is deleted.

Added logic for free-form mode to convert a morph polygon from Wipe to solid if the point count is not equal to 4.