VIX-2953 Polygon editor locked up while resizing
Removed some logic that was trying to assist when resizing to the edges.  It was only on the the middle top adorner so it was unbalanced anyway.

The other issue was with a Math.Round that was not doing what I thought it was doing.  Replaced it with Math.Ceiling.