VIX-2952 Fix the left to right drag select by accommodating for rotation.
The bounding box logic in the Custom prop is the non-rotated box. Override the base class method to allow for the rotation to be accounted for. This really needs a larger refactoring so the bounding box represents the true bounds of the prop. That is more extensive and will be left for future work.