VIX-2957 Fix multiple annoying bugs in the prop editor.
Provide a bit more leverage to the rotate handle so that there is more control on how much rotation occurs.

Fix a bug that could allow duplicate elements to be copied into a group when they appear in multiple groups at the same time.

Remove some of the handling of the initial directory in the open and save and let them default to the last paths as maintained by the OS to behave more like other applications.

Fix a bug in the way the resize adorner is rendered that would not always place it in the proper position over the entire selected set of lights. Added some logic to have the parent update itself when the adorner is added.