VIX-2953 - Fix scaling math
Original fix to (VIX-2953) was to take the ceiling of the scaled width and height. This had the potential to add some fractional slop to the drawing canvas. To fix this, I moved the logic to calculate the canvas size up so that the final scale factor can be based on the canvas size and the original display element size.

Once the scale factor was improved I noticed on my mega-tree that the scaling produced some floating point error which prevented the user from resizing the polygon. To account for the floating point error I added some rounding in the IsTransformValid method.

Finally I noticed there was an inconsistency on how single width display elements like arches were being scaled. This was updated to be consistent between the X and Y.