VIX-2987 Fix conflicts when rapid saves occur.
The saves when a user opens the Display setup or the Preview setup occur on a background thread to free up the UI. But the user is not prevented from opening the other setup while that is occurring. If they do that and close it before the other save completes, then conflicts can occur. Add logic to disable the buttons while the save is occurring. Add indicator in the progress bar on the admin that saving is in progress so the user is aware of what is going on.

Add wait cursors while saves are occurring.

Add the same blocking logic and progress indicator for the cancel path when the system is being reloaded.

Change the backup logic from a File move to a File copy to improve the performance. For some reason the move randomly is much slower. Copy seems to perform much faster and more consistent. This reduces the time the user is sitting waiting for the save to occur.

Fix the logic to move the Admin form back to the top after a system reload and utilize it on the reload from a cancel operation.