VIX-2999 Fix a threading issue in the DummyLighting Controller
Now that the profile modules are loading in a background thread, the DummyLighting was creating its form in the constructor instead of in the Start method and the form was not on the same thread as the message loop. Refactor the logic to create the form on the Start method correctly to solve the cross threading issue.