VIX-3011 Correct the MoveTo logic to account for Zoom level.
The move to logic was not scaling the movement offset back to the original zoom level that the prop coordinates are in and thus was multiplying the movement. This created erratic behavior the further from the original location the mouse was moving toward. Apply the reverse zoom logic to the move offset to get it into the same coordinate plane.