VIX-2878 Amend the fix from this original bug.
There are some cases where the default blank background will not be the same size as the view port. Change the logic to only look for props of zero size or ones that are off screen to the top left. There appears to be a bug that in some cases the no background previews do not have scroll bars when there are props outside the normal view port. This needs to be addressed as a bug in the future and then this can have logic added in to check the other bounds.