VIX-3016 Add size fields to the Custom Prop similar to other shapes.
Add the Top Left and Bottom Right fields.

Add logic to update them on any movement so you have live updates when moving a single prop.

Fix a bug that would reload the property viewer on mouse up even if the shape was already being edited. This was introducing an annoying flicker.