VIX-3018 Correct row zoom bug.
A seemingly do nothing line of code was removed in a previous change, however it was compensating for events being disabled after rows were resized. There is likely a better way to solve this in the future, but for now restored the line of code with some comments indicating its purpose.