VIX-2937 Initial attempt to use the sequence package import can fail when using the file selector.
The initial value of the import folder is uninitialized and the Path.GetDirectory method throws an path not of legal form exception even though it claims to accept null input. Added logic to test for empty or null and bypass the directory logic. Added the same to the output side to be safe even though it should be populated by default.