VIX-3045 Changes to allow additional vendor to be added to the prop vendor browser.
Gilbert has several props in their inventory that do not have models. This was unexpected when the initial work was done to support the browser. Add logic to handle the missing models.

Add validation logic to a the wiring model to determine if it is valid by looking to see if a link to it exists.

Add logic to hide the Model tab in the browser for a selected product that does not have a valid model as defined above.

Add logic to select the product tab any time the product selection changes so the user gets a proper view if the model is not valid.