VIX-3122 Fix the math logic for real this time.
Found other issues with the math logic converting back and forth when using more examples.

Added documentation as to how it works so it is easier to understand later.

Updated the text example of a Gamma curve formula to be more correct. A 2.8 Gamma is very aggressive. The formula also generated a small zero value that could keep items from being truly off at zero so I corrected the example and worded it in a way that hopefully makes it easier for a user to understand.