VIX-3130 Add offset marks feature.
Add logic to the MarkCollection model to offset the marks in it by a time span.

Add menu item to the collection context menu to Offset by time along with the logic to handle the event and prompt the user for the offset time.

Move appropriate icon to the resource area to be used on the context menu.

Fix the logic in the ruler to properly listen for Mark Collection changes so that it can redraw when changes occur.

Enhance the WPF MessageBoxService to provide for a numeric input option.