VIX-3142 Fix a bug in the profile selection screen that crashes Vixen.
Recent changes in VIX-988 introduced a bug that tries to clear the profile list box when a datasource is set. Changed the logic back to load the list items in a foreach after sorting to avoid using the datasource. This issue appears when the profile list is loaded a second time, most notably creating a new profile when others already exist.