VIX-3135 Add project and code from stand alone poc.
Renamed the import form in the editor as it has been added onto many times and retained the old original name referring to Audacity imports.

Incorporated the new logic to make the browser an option from the import form and then launch it when the user chooses.

Add views and models similar to the custom prop model browser, but with a more limited focus that pertained to the singing faces.