VIX-759 Error handling and file support enhancements.
Enhance the code to base the supported files off of the codecs that the system supports using CSCore.

Add additional error handling. This probably still is not reobust enough, but is better than the possibilities for a hard crash. The system itself is not well equipped to handle and audio player failure. That needs to be addressed in another effort.

Add a group box to the Options dialog for the audio latency to make clear it is for Audio and not Vixen itself.