VIX-3208 Improvements.
Improve the performance of the zig zag operation by changing it around to directly manipulate the order instead of using a temporary dictionary. Wrap it in begin and end updates to improve the UI rendering speed.

Add logic to support Ctrl-A selection of all the elements in the list view for those that want to apply the reverse or zig zag logic to all elements.