VIX-3209 Add patching order capabilities to the Pixel Grid.
Added UI controls to choose the corner the wiring starts from.

Added UI controls to select if Zig Zag should be applied when patching.

Added logic to the Patching order module to manage zig zag and standard patching for a given set of ordered nodes.

Added logic to the Pixel Gris setup screen to apply the patching order based on the users choices.