VIX-3255 Fix several issues with vendor url formatting.
Fix issue with the vendor url not opening on the singing faces project. Update to the new api to replace the deprecated form of opening a browser.

Fix issues with parsing the weblinks vendors provide in their custom props. Add logic to try and add the http:// when the vendor does not have that as part of the url.