VIX-3289 Several changes to improve the cleanup behavior in the display setup.
Add logic to ensure the highlighting is properly updated when a smart object is added to the preview with the creation template.

Add logic to update the highlighting and remove the dead linked nodes from the smart object when an element is deleted from the preview.

Add events to support knowing when elements are deleted from the element tree. Added a before delete, and an after delete. Added event args to support the type of edit. Not all paths use the extended event args. A larger improvement for another day.