VIX-3316 Fix issues related to VIX-3289 due to missing code commit.
Add logic to validate if the element still exists or not before unlinking to the delete routine. T

Add logic to trigger the remove action event args when nodes are deleted so they will properly update.

Fix a loop hole in the Reload method that could allow shapes to remained linked if just the nodeId is set. This is probably a corner case, but could happen.

Add logic to ensure when custom Props are added they are properly added to the NodeToPixel mapping and the highlight logic is called.

All of these changes were part of a final commit in the branch that I somehow lost in the original fix for VIX-3289 when I pushed the changes.