VIX-3345 Fix converter error for the MarkDecorator.
A warning started showing up in the logs after upgrading to .NET and the newest version of Catel on a property mismatch with the default converter. There was a cross usage of the interface vs concrete type that may have always been wrong. The fix was easy so getting out of the way.